Nessa branch foi desenvolvido todo o CRUD de informações pessoais dos usuÃários da aplicaação

+++ Back-end/techbridge/src/main/kotlin/com/techbridge/techbridge/repository/UsuarioRepository.kt
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UsuarioRepository : JpaRepository<Usuario, Int> {
+interface UsuarioRepository : JpaRepository<Usuario, Long> {
     fun findByEmail(email: String)
 }
